+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ _Date TBD_
 
 ### Features Added
 * Added EHR Certification ID rest service
+* Rearranged data for the corrective action plan reports/activity so there is just one activity event per CAP change.
 
 ### Bugs fixed
 * Fix parsing error when spreadsheet has only a single row.